rag-backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
asyncpg==0.29.0
qdrant-client==1.8.1
openai==1.3.5
python-dotenv==1.0.0
pyjwt[crypto]==2.8.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
better-ai==0.2.0